+++ cyberdrop_dl/crawlers/_chevereto.py
@@ -12,7 +12,6 @@ from cyberdrop_dl.utils.utilities import error_handling_wrapper
 
 if TYPE_CHECKING:
     from bs4 import BeautifulSoup
-    from yarl import URL
 
     from cyberdrop_dl.data_structures.url_objects import AbsoluteHttpURL, ScrapeItem
 from cyberdrop_dl.crawlers.crawler import SupportedPaths
@@ -146,7 +145,7 @@ class CheveretoCrawler(Crawler, is_abc=True):
         filename, ext = self.get_filename_and_ext(link.name)
         await self.handle_file(link, scrape_item, filename, ext)
 
-    async def get_embed_info(self, url: URL) -> tuple[str, URL]:
+    async def get_embed_info(self, url: AbsoluteHttpURL) -> tuple[str, AbsoluteHttpURL]:
         embed_url = self.PRIMARY_URL / "oembed" / ""
         embed_url = embed_url.with_query(url=str(url), format="json")
         async with self.request_limiter:
@@ -184,7 +183,7 @@ class CheveretoCrawler(Crawler, is_abc=True):
     video = partialmethod(_proccess_media_item, media_type=Media.VIDEO, selector=VIDEO_SELECTOR)
     image = partialmethod(_proccess_media_item, media_type=Media.IMAGE, selector=IMAGE_SELECTOR)
 
-    def get_canonical_url(self, url: URL, media_type: Media = Media.ALBUM) -> tuple[str, AbsoluteHttpURL]:
+    def get_canonical_url(self, url: AbsoluteHttpURL, media_type: Media = Media.ALBUM) -> tuple[str, AbsoluteHttpURL]:
         """Returns the id and canonical URL from a given item (album, image or video)."""
         search_parts = ALBUM_PARTS
         if media_type == Media.IMAGE: