+++ cyberdrop_dl/crawlers/_forum.py
@@ -141,7 +141,6 @@ class ThreadProtocol(Protocol):
     def url(self) -> AbsoluteHttpURL: ...
 
 
-# quotes_2 : embed fauxBlockLink
 class MessageBoardCrawler(Crawler, is_abc=True):
     """Base crawler for every MessageBoard.
 
@@ -229,6 +228,9 @@ class MessageBoardCrawler(Crawler, is_abc=True):
 
     async def fetch_thread(self, scrape_item: ScrapeItem) -> None:
         thread_part_index = len(self.PRIMARY_URL.parts)
+        # https://github.com/jbsparrow/CyberDropDownloader/issues/1165#issuecomment-3086739753
+        if self.PRIMARY_URL.parts[-1] == "":
+            thread_part_index -= 1
         match scrape_item.url.parts[thread_part_index:]:
             case [thread_part, thread_name_and_id, *_] if thread_part in self.THREAD_PART_NAMES:
                 self.check_thread_recursion(scrape_item)
@@ -502,7 +504,7 @@ class HTMLMessageBoardCrawler(MessageBoardCrawler, is_abc=True):
             max_children_error = e
 
         if seen:
-            self.log(f"post #{post.id} {stats = }")
+            self.log(f"[{self.FOLDER_DOMAIN}] post #{post.id} {stats = }")
         if duplicates:
             msg = f"Found duplicate links in post {scrape_item.url}. Selectors are too generic: {duplicates}"
             self.log(msg, bug=True)