+++ cyberdrop_dl/crawlers/camwhores_dot_tv.py
@@ -3,8 +3,9 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, ClassVar
 
 from cyberdrop_dl.crawlers._kvs import KernelVideoSharingCrawler, Video
+from cyberdrop_dl.crawlers.porntrex import PorntrexCrawler
 from cyberdrop_dl.data_structures.url_objects import AbsoluteHttpURL, ScrapeItem
-from cyberdrop_dl.utils.utilities import get_og_properties
+from cyberdrop_dl.utils import css, open_graph
 
 if TYPE_CHECKING:
     from bs4 import BeautifulSoup
@@ -13,6 +14,7 @@ if TYPE_CHECKING:
 
 
 PRIMARY_URL = AbsoluteHttpURL("https://www.camwhores.tv")
+LAST_PAGE_SELECTOR = "div.pagination-holder li.page"
 
 
 class CamwhoresTVCrawler(KernelVideoSharingCrawler):
@@ -35,7 +37,7 @@ class CamwhoresTVCrawler(KernelVideoSharingCrawler):
 
     def get_video_info(self, soup: BeautifulSoup) -> Video:
         video = super().get_video_info(soup)
-        return video._replace(title=get_og_properties(soup).title)
+        return video._replace(title=open_graph.title(soup))
 
     def parse_url(self, link_str: str, relative_to: AbsoluteHttpURL | None = None, *_) -> AbsoluteHttpURL:
         return super().parse_url(link_str, relative_to, trim=False)
@@ -49,3 +51,18 @@ class CamwhoresTVCrawler(KernelVideoSharingCrawler):
 
     async def album(self, scrape_item: ScrapeItem) -> None:
         raise NotImplementedError
+
+    async def iter_videos(self, scrape_item: ScrapeItem, video_category: str = "") -> None:
+        url = scrape_item.url / video_category if video_category else scrape_item.url
+        await super().iter_videos(scrape_item, video_category)
+        async with self.request_limiter:
+            soup: BeautifulSoup = await self.client.get_soup(self.DOMAIN, url)
+
+        last_page = int(css.get_text(soup.select(LAST_PAGE_SELECTOR)[-1]))
+        # TODO: Porntrex also uses KVS. Make the KVS crawler handle it by default
+        await PorntrexCrawler.proccess_additional_pages(
+            self,  # type: ignore[ArgumentType]
+            scrape_item,
+            last_page,
+            block_id="list_videos_common_videos_list",
+        )