+++ cyberdrop_dl/crawlers/eporner.py
@@ -1,16 +1,17 @@
 from __future__ import annotations
 
+import dataclasses
+import json
 from typing import TYPE_CHECKING, ClassVar, NamedTuple
 
 from cyberdrop_dl.crawlers.crawler import Crawler, SupportedPaths
 from cyberdrop_dl.data_structures.url_objects import AbsoluteHttpURL, ScrapeItem
 from cyberdrop_dl.exceptions import ScrapeError
-from cyberdrop_dl.utils import css, javascript
+from cyberdrop_dl.utils import css
 from cyberdrop_dl.utils.utilities import error_handling_wrapper
 
 if TYPE_CHECKING:
     from bs4 import BeautifulSoup, Tag
-    from yarl import URL
 
     from cyberdrop_dl.data_structures.url_objects import ScrapeItem
 
@@ -41,14 +42,21 @@ PROFILE_URL_PARTS = {
 }
 
 
-class VideoInfo(NamedTuple):
+@dataclasses.dataclass(frozen=True, slots=True)
+class Video:
+    title: str
+    date: str
+    best_src: VideoSource
+
+
+class VideoSource(NamedTuple):
     codec: str  # av1 > h264
     resolution: str
     size: str
-    link_str: str
+    url: str
 
     @classmethod
-    def from_tag(cls, tag: Tag) -> VideoInfo:
+    def from_tag(cls, tag: Tag) -> VideoSource:
         link_str: str = css.get_attr(tag, "href")
         name = tag.get_text(strip=True).removeprefix("Download")
         details = name.split("(", 1)[1].removesuffix(")").split(",")
@@ -74,7 +82,7 @@ class EpornerCrawler(Crawler):
     NEXT_PAGE_SELECTOR: ClassVar[str] = _SELECTORS.NEXT_PAGE
 
     async def fetch(self, scrape_item: ScrapeItem) -> None:
-        if get_video_id(scrape_item.url):
+        if _get_video_id(scrape_item.url):
             return await self.video(scrape_item)
         if any(p in scrape_item.url.parts for p in ("cat", "channel", "search", "pornstar")):
             return await self.playlist(scrape_item)
@@ -153,17 +161,14 @@ class EpornerCrawler(Crawler):
             soup: BeautifulSoup = await self.client.get_soup(self.DOMAIN, scrape_item.url)
 
         scrape_item.url = canonical_url
-        img = soup.select_one(_SELECTORS.PHOTO)
-        if not img:
-            raise ScrapeError(422)
-        link_str: str = css.get_attr(img, "href")
+        link_str = css.select_one_get_attr(soup, _SELECTORS.PHOTO, "href")
         link = self.parse_url(link_str)
         filename, ext = self.get_filename_and_ext(link.name)
         await self.handle_file(link, scrape_item, filename, ext)
 
     @error_handling_wrapper
     async def video(self, scrape_item: ScrapeItem) -> None:
-        video_id = get_video_id(scrape_item.url)
+        video_id = _get_video_id(scrape_item.url)
         canonical_url = PRIMARY_URL / f"video-{video_id}"
         if await self.check_complete_from_referer(canonical_url):
             return
@@ -178,56 +183,44 @@ class EpornerCrawler(Crawler):
             raise ScrapeError(410)
 
         scrape_item.url = canonical_url
-        info_dict = get_info_dict(soup)
-        resolution, link_str = get_best_quality(soup)
-        if not link_str:
-            raise ScrapeError(422)
-
-        link = self.parse_url(link_str)
-        scrape_item.possible_datetime = self.parse_date(info_dict["uploadDate"])
+        video = _parse_video(soup)
+        link = self.parse_url(video.best_src.url)
+        scrape_item.possible_datetime = self.parse_date(video.date)
         filename, ext = self.get_filename_and_ext(link.name)
-        if ext == ".m3u8":
-            raise ScrapeError(422)
-        custom_filename = self.create_custom_filename(info_dict["name"], ext, file_id=video_id, resolution=resolution)
+        custom_filename = self.create_custom_filename(
+            video.title, ext, file_id=video_id, resolution=video.best_src.resolution
+        )
         await self.handle_file(link, scrape_item, filename, ext, custom_filename=custom_filename)
 
 
-def get_available_resolutions(soup: BeautifulSoup) -> list[VideoInfo]:
-    downloads = soup.select_one(_SELECTORS.DOWNLOADS)
-    assert downloads
+def _get_available_sources(soup: BeautifulSoup) -> list[VideoSource]:
+    downloads = css.select_one(soup, _SELECTORS.DOWNLOADS)
     formats = downloads.select(_SELECTORS.H264)
     if ALLOW_AV1:
         formats.extend(downloads.select(_SELECTORS.AV1))
-    return [VideoInfo.from_tag(tag) for tag in formats]
+    return [VideoSource.from_tag(tag) for tag in formats]
 
 
-def get_best_quality(soup: BeautifulSoup) -> tuple[str, str]:
-    """Returns name and URL of the best available quality.
-
-    Returns URL as `str`"""
-    formats = get_available_resolutions(soup)
-    best_format = formats[0]
-    formats_dict: dict[str, list[VideoInfo]] = {}
+def _get_best_src(soup: BeautifulSoup) -> VideoSource:
+    formats = _get_available_sources(soup)
     for res in RESOLUTIONS:
-        formats_dict[res] = sorted(f for f in formats if f.resolution == res)
-
-    for res in RESOLUTIONS:
-        available_formats = formats_dict.get(res)
-        if available_formats:
-            best_format = available_formats[0]
-            break
-
-    return best_format.resolution, best_format.link_str
+        sorted_formats = sorted(f for f in formats if f.resolution == res and not f.url.endswith(".m3u8"))
+        if sorted_formats:
+            return sorted_formats[0]
+    raise ScrapeError(422, message="Unable to parse video formats")
 
 
-def get_info_dict(soup: BeautifulSoup) -> dict:
-    info_js_script = css.select_one_get_text(soup, _SELECTORS.DATE_JS)
-    info_dict: dict = javascript.parse_json_to_dict(info_js_script, use_regex=False)
-    javascript.clean_dict(info_dict)
-    return info_dict
+def _parse_video(soup: BeautifulSoup) -> Video:
+    ld_json = css.select_one_get_text(soup, _SELECTORS.DATE_JS)
+    video_data = json.loads(ld_json)
+    return Video(
+        title=video_data["name"],
+        date=video_data["uploadDate"],
+        best_src=_get_best_src(soup),
+    )
 
 
-def get_video_id(url: URL) -> str:
+def _get_video_id(url: AbsoluteHttpURL) -> str:
     if "video-" in url.parts[1]:
         return url.parts[1].rsplit("-", 1)[1]
     if any(p in url.parts for p in ("hd-porn", "embed")) and len(url.parts) > 2: