+++ cyberdrop_dl/crawlers/imx_to.py
@@ -9,8 +9,6 @@ from cyberdrop_dl.utils import css
 from cyberdrop_dl.utils.utilities import error_handling_wrapper
 
 if TYPE_CHECKING:
-    from yarl import URL
-
     from cyberdrop_dl.data_structures.url_objects import ScrapeItem
 
 
@@ -63,5 +61,5 @@ class ImxToCrawler(Crawler):
         return PRIMARY_URL / get_image_id(url)
 
 
-def get_image_id(url: URL) -> str:
+def get_image_id(url: AbsoluteHttpURL) -> str:
     return Path(url.name).stem