+++ cyberdrop_dl/crawlers/rule34xxx.py
@@ -4,7 +4,6 @@ from typing import TYPE_CHECKING, ClassVar
 
 from cyberdrop_dl.crawlers.crawler import Crawler, SupportedPaths
 from cyberdrop_dl.data_structures.url_objects import AbsoluteHttpURL
-from cyberdrop_dl.exceptions import ScrapeError
 from cyberdrop_dl.utils import css
 from cyberdrop_dl.utils.utilities import error_handling_wrapper
 
@@ -21,9 +20,10 @@ class Selector:
     IMAGE = "img[id=image]"
     VIDEO = "video source"
     DATE = "li:contains('Posted: ')"
+    IMAGE_OR_VIDEO = f"{IMAGE}, {VIDEO}"
 
 
-_SELECTOR = Selector()
+_SELECTORS = Selector()
 
 
 class Rule34XXXCrawler(Crawler):
@@ -34,7 +34,7 @@ class Rule34XXXCrawler(Crawler):
     FOLDER_DOMAIN: ClassVar[str] = "Rule34XXX"
 
     async def async_startup(self) -> None:
-        self.set_cookies()
+        self.update_cookies({"resize-original": "1"})
 
     async def fetch(self, scrape_item: ScrapeItem) -> None:
         if scrape_item.url.query.get("tags"):
@@ -52,7 +52,7 @@ class Rule34XXXCrawler(Crawler):
                 title = self.create_title(title_portion)
                 scrape_item.setup_as_album(title)
 
-            for _, new_scrape_item in self.iter_children(scrape_item, soup, _SELECTOR.CONTENT):
+            for _, new_scrape_item in self.iter_children(scrape_item, soup, _SELECTORS.CONTENT):
                 self.manager.task_group.create_task(self.run(new_scrape_item))
 
     @error_handling_wrapper
@@ -60,18 +60,9 @@ class Rule34XXXCrawler(Crawler):
         async with self.request_limiter:
             soup: BeautifulSoup = await self.client.get_soup(self.DOMAIN, scrape_item.url)
 
-        media_tag = soup.select_one(_SELECTOR.IMAGE) or soup.select_one(_SELECTOR.VIDEO)
-        if not media_tag:
-            raise ScrapeError(422)
-
-        if date_tag := soup.select_one(_SELECTOR.DATE):
-            date_str = date_tag.get_text(strip=True).removeprefix("Posted: ")
-            scrape_item.possible_datetime = self.parse_date(date_str)
-        link_str: str = css.get_attr(media_tag, "src")
+        date_str = css.select_one_get_text(soup, _SELECTORS.DATE).removeprefix("Posted: ")
+        scrape_item.possible_datetime = self.parse_date(date_str)
+        link_str = css.select_one_get_attr(soup, _SELECTORS.IMAGE_OR_VIDEO, "src")
         link = self.parse_url(link_str)
         filename, ext = self.get_filename_and_ext(link.name)
         await self.handle_file(link, scrape_item, filename, ext)
-
-    def set_cookies(self) -> None:
-        cookies = {"resize-original": "1"}
-        self.update_cookies(cookies)