+++ cyberdrop_dl/crawlers/spankbang.py
@@ -1,15 +1,15 @@
 from __future__ import annotations
 
 import itertools
+import json
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, Any, ClassVar, NamedTuple
+from typing import TYPE_CHECKING, ClassVar, NamedTuple
 
 from cyberdrop_dl.crawlers.crawler import Crawler, SupportedPaths
 from cyberdrop_dl.data_structures.url_objects import AbsoluteHttpURL, ScrapeItem
 from cyberdrop_dl.exceptions import ScrapeError
 from cyberdrop_dl.utils import css, javascript
-from cyberdrop_dl.utils.logger import log_debug
-from cyberdrop_dl.utils.utilities import error_handling_wrapper
+from cyberdrop_dl.utils.utilities import error_handling_wrapper, get_text_between
 
 if TYPE_CHECKING:
     from bs4 import BeautifulSoup
@@ -24,8 +24,8 @@ RESOLUTIONS = ["4k", "2160p", "1440p", "1080p", "720p", "480p", "360p", "240p"] 
 VIDEO_REMOVED_SELECTOR = "[id='video_removed'], [class*='video_removed']"
 VIDEOS_SELECTOR = "div.video-list > div.video-item > a"
 
-JS_SELECTOR = "main.main-container > script:contains('var stream_data')"
-EXTENDED_JS_SELECTOR = "main.main-container > script:contains('uploadDate')"
+JS_STREAM_DATA_SELECTOR = "main.main-container > script:contains('var stream_data')"
+JS_VIDEO_INFO_SELECTOR = "main.main-container > script:contains('uploadDate')"
 
 
 @dataclass(frozen=True)
@@ -48,8 +48,12 @@ class Format(NamedTuple):
     link_str: str
 
 
-## TODO: convert to global dataclass with constructor from dict to use in multiple crawlers
-class VideoInfo(dict): ...
+@dataclass(frozen=True, slots=True, kw_only=True)
+class Video:
+    id: str
+    title: str
+    date: str
+    best_format: Format
 
 
 class SpankBangCrawler(Crawler):
@@ -67,10 +71,10 @@ class SpankBangCrawler(Crawler):
     FOLDER_DOMAIN: ClassVar[str] = "SpankBang"
 
     async def async_startup(self) -> None:
-        self.set_cookies()
+        self.update_cookies({"country": "US", "age_pass": 1})
 
     async def fetch(self, scrape_item: ScrapeItem) -> None:
-        if is_playlist(scrape_item.url):
+        if _is_playlist(scrape_item.url):
             return await self.playlist(scrape_item)
         if any(p in scrape_item.url.parts for p in ("video", "play", "embed", "playlist")):
             return await self.video(scrape_item)
@@ -118,58 +122,44 @@ class SpankBangCrawler(Crawler):
             soup: BeautifulSoup = await self.client.get_soup_cffi(self.DOMAIN, scrape_item.url)
 
         was_removed = soup.select_one(VIDEO_REMOVED_SELECTOR)
-        if was_removed or "This video is no longer available" in str(soup):
+        if was_removed or "This video is no longer available" in soup.text:
             raise ScrapeError(410)
 
-        info = get_info_dict(soup)
-        canonical_url = PRIMARY_URL / info["video_id"] / "video"
+        video = _parse_video(soup)
+        canonical_url = PRIMARY_URL / video.id / "video"
         if await self.check_complete_from_referer(canonical_url):
             return
         scrape_item.url = canonical_url
-
-        video_format = get_best_quality(info)
-        if not video_format:
-            raise ScrapeError(422)
-        resolution, link_str = video_format
-        scrape_item.possible_datetime = self.parse_iso_date(info["uploadDate"])
+        resolution, link_str = video.best_format
+        scrape_item.possible_datetime = self.parse_iso_date(video.date)
         link = self.parse_url(link_str)
         filename, ext = self.get_filename_and_ext(link.name)
-        custom_filename = self.create_custom_filename(info["title"], ext, file_id=video_id, resolution=resolution)
+        custom_filename = self.create_custom_filename(video.title, ext, file_id=video_id, resolution=resolution)
         await self.handle_file(link, scrape_item, filename, ext, custom_filename=custom_filename)
 
-    def set_cookies(self) -> None:
-        cookies = {"country": "US", "age_pass": 1}
-        self.update_cookies(cookies)
 
-
-def get_info_dict(soup: BeautifulSoup) -> VideoInfo:
-    info_js_script_text = css.select_one_get_text(soup, JS_SELECTOR)
-    extended_info_js_script_text = css.select_one_get_text(soup, EXTENDED_JS_SELECTOR)
-
+def _parse_video(soup: BeautifulSoup) -> Video:
     title_tag = css.select_one(soup, "div#video h1")
-    title: str = css.get_attr_or_none(title_tag, "title") or css.get_text(title_tag).replace("\n", "")
-    del soup
-    info: dict[str, Any] = javascript.parse_js_vars(info_js_script_text)
-    extended_info_dict: dict = javascript.parse_json_to_dict(extended_info_js_script_text)
+    stream_js_text = css.select_one_get_text(soup, JS_STREAM_DATA_SELECTOR)
+    js_text = css.select_one_get_text(soup, JS_VIDEO_INFO_SELECTOR)
+    video_data = json.loads(js_text)
+    stream_data = javascript.parse_obj(get_text_between(stream_js_text, "stream_data = ", ";"))
+    embed_url = AbsoluteHttpURL(video_data["embedUrl"])
+    return Video(
+        id=embed_url.parts[1],
+        title=css.get_attr_or_none(title_tag, "title") or css.get_text(title_tag),
+        date=video_data["uploadDate"],
+        best_format=_get_best_quality(stream_data),
+    )
 
-    info["title"] = title.strip()
-    info = info | extended_info_dict
-    embed_url = AbsoluteHttpURL(info["embedUrl"])
-    info["video_id"] = embed_url.parts[1]
-    javascript.clean_dict(info, "stream_data")
-    log_debug(info)
-    return VideoInfo(**info)
 
-
-def get_best_quality(info_dict: dict) -> Format:
+def _get_best_quality(stream_data: dict[str, list[str]]) -> Format:
     """Returns name and URL of the best available quality."""
-    qualities: dict = info_dict["stream_data"]
     for res in RESOLUTIONS:
-        value = qualities.get(res)
-        if value:
+        if value := stream_data.get(res):
             return Format(res, value[-1])
-    return Format(DEFAULT_QUALITY, qualities[DEFAULT_QUALITY])
+    raise ScrapeError(422, message="Unable to get download link")
 
 
-def is_playlist(url: AbsoluteHttpURL) -> bool:
+def _is_playlist(url: AbsoluteHttpURL) -> bool:
     return "playlist" in url.parts and "-" not in url.parts[1]