+++ cyberdrop_dl/utils/logger.py
@@ -85,17 +85,6 @@ class BareQueueHandler(QueueHandler):
         return record
 
 
-class TextualLogQueueHandler(QueueHandler):
-    """Auto queue and format the log record as a rich text before sending it to the queue"""
-
-    def __init__(self, manager: Manager):
-        manager.textual_log_queue = q = queue.Queue()
-        super().__init__(q)
-
-    def prepare(self, record: logging.LogRecord) -> Text:
-        return create_rich_log_msg(record.getMessage(), record.levelno)
-
-
 class QueuedLogger:
     """A helper class to setup a queue handler + listener."""
 
@@ -233,10 +222,9 @@ def log(
 ) -> None:
     """Simple logging function."""
     msg = prefix + process_log_msg(message)
-    logger.log(level, msg, **kwargs)
     log_debug(msg, level, **kwargs)
     if bug:
-        msg = msg.rstrip() + f". Please file a bug report at {NEW_ISSUE_URL}"
+        msg = msg.rstrip() + f". Please open a bug report at {NEW_ISSUE_URL}"
     logger.log(level, msg, **kwargs)