+++ CHANGELOG.md
@@ -22,6 +22,18 @@ All notable changes to this project will be documented here. For more details, v
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [7.2.1] - 2025-07-25
+
+### Changed
+
+- Add back crawler name when logging posts stats (Xenforo)
+
+### Fixed
+
+- Messages being logged twice
+- Tag pagination (camwhores)
+- `Generator didn't yield` error when CDL finds an invalid URL
+
 ## [7.2.0] - 2025-07-17
 
 ### Added